fsck.ocfs2: Silence build errors on x86_64
Silence the following build errors on x86_64.

extent.c: In function ‘o2fsck_check_extent_rec’:
extent.c:406: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 6 has type ‘__u64’
pass1b.c: In function ‘process_dup_clusters’:
pass1b.c:387: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 5 has type ‘__u64’
pass1b.c:392: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 5 has type ‘__u64’
pass1b.c:401: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 5 has type ‘__u64’
pass1b.c: In function ‘process_xattr_header’:
pass1b.c:506: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 4 has type ‘__u64’
pass1b.c: In function ‘process_one_bucket_list’:
pass1b.c:546: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 5 has type ‘__u64’
pass1b.c: In function ‘process_xattr_tree’:
pass1b.c:608: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 4 has type ‘__u64’
pass1b.c: In function ‘process_xattr_block’:
pass1b.c:627: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 4 has type ‘__u64’
pass2.c: In function ‘fix_dir_trailer’:
pass2.c:239: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 8 has type ‘__u64’
pass2.c:250: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 8 has type ‘__u64’
pass5.c: In function ‘o2fsck_check_data_blk’:
pass5.c:243: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 7 has type ‘__u64’
pass5.c:243: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 8 has type ‘__u64’
pass5.c:243: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 9 has type ‘__u64’
pass5.c:243: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 10 has type ‘__u64’
pass5.c:243: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 11 has type ‘__u64’
pass5.c:243: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 12 has type ‘__u64’
pass5.c:243: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 13 has type ‘__u64’
pass5.c:243: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 14 has type ‘__u64’
pass5.c:261: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 7 has type ‘__u64’
pass5.c:261: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 8 has type ‘__u64’
pass5.c:261: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 9 has type ‘__u64’
pass5.c:261: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 10 has type ‘__u64’
pass5.c:261: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 11 has type ‘__u64’
pass5.c:261: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 12 has type ‘__u64’
pass5.c:261: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 13 has type ‘__u64’
pass5.c:261: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 14 has type ‘__u64’
pass5.c:280: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 7 has type ‘__u64’
pass5.c:280: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 8 has type ‘__u64’
pass5.c:280: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 9 has type ‘__u64’
pass5.c:280: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 10 has type ‘__u64’
util.c: In function ‘o2fsck_print_resource_track’:
util.c:325: warning: format ‘%u’ expects type ‘unsigned int’, but argument 2 has type ‘size_t’
util.c:330: warning: format ‘%llu’ expects type ‘long long unsigned int’, but argument 2 has type ‘uint64_t’
util.c:330: warning: format ‘%llu’ expects type ‘long long unsigned int’, but argument 3 has type ‘uint64_t’
util.c:330: warning: format ‘%llu’ expects type ‘long long unsigned int’, but argument 4 has type ‘uint64_t’
xattr.c: In function ‘check_xattr_entry’:
xattr.c:287: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 8 has type ‘__u64’

+++ fsck.ocfs2/extent.c
@@ -407,8 +407,8 @@ errcode_t o2fsck_check_extent_rec(o2fsck_state *ost,
 		   "The extent record for cluster offset %"PRIu32" at block "
 		   "%"PRIu64" in inode %"PRIu64" has the REFCOUNTED flag set, "
 		   "while it shouldn't have that flag. "
-		   "Clear the REFCOUNTED flag?", er->e_cpos, er->e_blkno,
-		   (uint64_t)di->i_blkno)) {
+		   "Clear the REFCOUNTED flag?", er->e_cpos,
+		   (uint64_t)er->e_blkno, (uint64_t)di->i_blkno)) {
 		er->e_flags &= ~OCFS2_EXT_REFCOUNTED;
 		*changed = 1;
 	}
